refactor: Moved url functions to new file
