+++ src/sdk/helpers/requests.ts
@@ -1,22 +1,6 @@
 /* eslint-disable camelcase */
-import { Name, PublicKey } from '@wharfkit/antelope';
-import { IDContract } from '../services/blockchain/contracts/IDContract';
-import { KeyManager, KeyManagerLevel } from '../storage/keymanager';
-import { createStorage, PersistentStorageClean, StorageFactory, STORAGE_NAMESPACE } from '../storage/storage';
-import { User } from '../controllers/user';
-import { createKeyManagerSigner } from '../services/blockchain/eosio/transaction';
-import { SdkError, SdkErrors, throwError } from '../util/errors';
-import { App, AppStatus } from '../controllers/app';
-import { TonomyUsername } from '../util/username';
+import { SdkErrors, throwError } from '../util/errors';
 import { LoginRequest, TonomyRequest } from '../util/request';
-import { LoginRequestResponseMessage, LoginRequestsMessagePayload } from '../services/communication/message';
-import { LoginRequestResponseMessagePayload } from '../services/communication/message';
-import { base64UrlToObj, objToBase64Url } from '../util/base64';
-import { getSettings } from '../util/settings';
-import { DID, URL as URLtype } from '../util/ssi/types';
-import { Issuer } from '@tonomy/did-jwt-vc';
-import { ES256KSigner, JsKeyManager, createVCSigner, generateRandomKeyPair } from '..';
-import { createJWK, toDid } from '../util/ssi/did-jwk';
 import { DataSharingRequest } from '../util';
 
 /**
