+++ src/sdk/helpers/userApps.ts
@@ -7,16 +7,13 @@ import { User } from '../controllers/user';
 import { createKeyManagerSigner } from '../services/blockchain/eosio/transaction';
 import { SdkError, SdkErrors, throwError } from '../util/errors';
 import { App, AppStatus } from '../controllers/app';
-import { TonomyUsername } from '../util/username';
 import { LoginRequest, TonomyRequest } from '../util/request';
-import { LoginRequestResponseMessage, LoginRequestsMessagePayload } from '../services/communication/message';
+import { LoginRequestResponseMessage } from '../services/communication/message';
 import { LoginRequestResponseMessagePayload } from '../services/communication/message';
-import { base64UrlToObj, objToBase64Url } from '../util/base64';
+import { objToBase64Url } from '../util/base64';
 import { getSettings } from '../util/settings';
 import { DID, URL as URLtype } from '../util/ssi/types';
 import { Issuer } from '@tonomy/did-jwt-vc';
-import { ES256KSigner, JsKeyManager, createVCSigner, generateRandomKeyPair } from '..';
-import { createJWK, toDid } from '../util/ssi/did-jwk';
 import { DataSharingRequest } from '../util';
 import { verifyRequests } from './requests';
 
@@ -226,120 +223,6 @@ export class UserApps {
         }
 
         return response;
-    }
-
-    /**
-     * Extracts the TonomyRequests from the URL
-     *
-     * @returns {LoginRequestsMessagePayload} the requests, username and accountName
-     */
-    static getLoginRequestFromUrl(): LoginRequestsMessagePayload {
-        const params = new URLSearchParams(window.location.search);
-
-        const base64UrlPayload = params.get('payload');
-
-        if (!base64UrlPayload) throwError("payload parameter doesn't exist", SdkErrors.MissingParams);
-
-        // get unparsed LoginRequestsMessagePayload object
-        const unparsedLoginRequestMessagePayload = base64UrlToObj(base64UrlPayload);
-
-        if (!unparsedLoginRequestMessagePayload.requests)
-            throwError('No requests found in payload', SdkErrors.MissingParams);
-
-        const unparsedRequestStrings = unparsedLoginRequestMessagePayload.requests.filter(
-            (request: string) => request !== null
-        );
-
-        if (!unparsedRequestStrings) throwError('No requests found in payload', SdkErrors.MissingParams);
-
-        const requests = unparsedRequestStrings.map((request: string) => {
-            const tonomyRequest = new TonomyRequest(request);
-
-            if (tonomyRequest.getType() === LoginRequest.getType()) {
-                return new LoginRequest(tonomyRequest);
-            } else if (tonomyRequest.getType() === DataSharingRequest.getType()) {
-                return new DataSharingRequest(tonomyRequest);
-            } else {
-                throwError('Invalid TonomyRequest Type', SdkErrors.InvalidRequestType);
-            }
-        });
-
-        return { requests };
-    }
-
-    /**
-     * Extracts the login requests, username and accountName from the URL
-     *
-     * @returns {LoginRequestResponseMessagePayload} the requests, username and accountName
-     */
-    static getLoginRequestResponseFromUrl(): LoginRequestResponseMessagePayload {
-        const params = new URLSearchParams(window.location.search);
-
-        const base64UrlPayload = params.get('payload');
-
-        if (!base64UrlPayload) throwError("payload parameter doesn't exists", SdkErrors.MissingParams);
-
-        const parsedPayload = base64UrlToObj(base64UrlPayload);
-
-        if (parsedPayload.success !== true && parsedPayload.success !== false)
-            throwError("success parameter doesn't exists", SdkErrors.MissingParams);
-
-        const { requests } = this.getLoginRequestFromUrl();
-
-        if (parsedPayload.success) {
-            if (!parsedPayload.accountName) throwError("accountName parameter doesn't exists", SdkErrors.MissingParams);
-            const res: LoginRequestResponseMessagePayload = {
-                success: true,
-                requests,
-                accountName: Name.from(parsedPayload.accountName),
-            };
-
-            if (parsedPayload.username) res.username = new TonomyUsername(parsedPayload.username);
-            return res;
-        } else {
-            if (!parsedPayload.error) throwError("error parameter doesn't exists", SdkErrors.MissingParams);
-            return { success: false, requests, error: parsedPayload.error };
-        }
-    }
-
-    /**
-     * Verifies the TonomyRequest received in the URL were successfully authorized by Tonomy ID
-     *
-     * @description should be called in the callback page of the SSO Login website
-     *
-     * @returns {Promise<TonomyRequest>} - the verified TonomyRequest
-     */
-    static async onRedirectLogin(): Promise<TonomyRequest> {
-        const { requests } = this.getLoginRequestFromUrl();
-
-        const verifiedRequests = await verifyRequests(requests);
-
-        const docReferrer = document.referrer;
-
-        if (!docReferrer) throwError('No referrer found', SdkErrors.ReferrerEmpty);
-
-        const referrer = new URL(docReferrer);
-
-        const myRequest = verifiedRequests.find((request) => {
-            if (request.getType() === LoginRequest.getType()) {
-                const loginRequest = request.getPayload();
-
-                return loginRequest.origin === referrer.origin;
-            }
-
-            return false;
-        });
-
-        if (!myRequest) {
-            const msg =
-                `No origins from: ${verifiedRequests.find(
-                    (r) => r.getType() === LoginRequest.getType() && r.getPayload().origin
-                )} ` + `match referrer: ${referrer.origin}`;
-
-            throwError(msg, SdkErrors.WrongOrigin);
-        }
-
-        return myRequest;
     }
 
     /**
